requirements.txt
aiohttp
py-cord
google-generativeai
python-dotenv

